add playback functionality

+++ reprounzip/unpackers/dj.py
@@ -13,7 +13,7 @@ from collections import deque
 import os
 import tarfile
 from reprounzip import signals
-from reprounzip.unpackers.docker import docker_setup, docker_run
+from reprounzip.unpackers.docker import docker_setup, docker_run, read_dict
 
 logger = logging.getLogger('reprounzip.dj')
 logger.setLevel(10)
@@ -24,51 +24,33 @@ if len(logger.handlers) < 1:
 class SubprocessManager(object):
 
     def __init__(self):
-        self.procs = []
+        self.running = []
 
-    def register(self, proc):
-        self.procs.append(proc)
+    def register(self, stopable):
+        self.running.append(stopable)
 
     def shutdown(self):
-        for p in self.procs:
-            p.kill()
-        logger.debug("all procs shut down")
-
-subprocess_manager = SubprocessManager()
-
-def shutdown(sig, frame):
-    subprocess_manager.shutdown()
-    sys.exit(0)
+        for r in self.running:
+            print(repr(r))
+            try:
+                r.stop()
+            except docker.errors.NotFound:
+                pass
+        logger.debug("all jobs stopped")
 
 
-# Wraps the wayback record service
-class Recorder(object):
+# Wraps the wayback service
+# note: playback functionality moved to docker
+class Wayback(object):
+    PORT = 8080
 
-    def __init__(self, root_dir, quiet=False):
-        self.root_dir = root_dir
-        self.port = 8080
-        self.quiet = quiet
-
-    def start(self):
-        args = {}
-        if self.quiet:
-            args['stdout'] = subprocess.DEVNULL
-            args['stderr'] = subprocess.DEVNULL
-
-        try:
-            print(self.root_dir)
-            self.proc = subprocess.Popen(['wayback', '--record', '--live', '-a', '--auto-interval', '5', '-d', self.root_dir], **args)
-
-        except Exception as e:
-            logger.debug(e)
-            logger.warn("Wayback service failed to start")
-            raise e
-
+    @staticmethod
+    def wait_for_service(port):
         tries = 10
         success = False
         while (tries > 0):
             try:
-                r = requests.get("http://localhost:{}".format(self.port))
+                r = requests.get("http://localhost:{}".format(Wayback.PORT))
                 if r.status_code == 200:
                     success = True
                     break
@@ -80,45 +62,111 @@ class Recorder(object):
                 time.sleep(5)
 
         if success:
-            logger.info("Recorder successfully started")
-        else:
-            raise Exception("Recorder failed to start")
+            return
+        raise Exception("Wayback failed to start correctly")
 
-        return self
 
-    def kill(self):
-        res = self.proc.kill()
-        return res
+    @classmethod
+    def new_recorder(cls, root_dir, args):
+        popen_args = [
+            '--record',
+            '--live',
+            '-a',
+            '--auto-interval',
+            '5',
+            '-d',
+            root_dir]
+        return cls(popen_args, args)
+
+    @classmethod
+    def new_replayer(cls, root_dir, args=None):
+        popen_args = [
+            '--proxy', 'warc-data', '-d', root_dir]
+        return cls(popen_args, args)
+
+    def __init__(self, proc_args, args=None):
+        self.proc = None
+        self.proc_args = proc_args
+        self.output_args = {}
+        if args and args.quiet:
+            self.output_args['stdout'] = subprocess.DEVNULL
+            self.output_args['stderr'] = subprocess.DEVNULL
+
+    def stop(self):
+        try:
+            self.proc.kill()
+        except AttributeError:
+            pass
+
+
+    def start(self):
+        try:
+            proc_args = ['wayback', '-p', str(Wayback.PORT)] + self.proc_args
+            logger.debug(proc_args)
+            self.proc = subprocess.Popen(proc_args, **self.output_args)
+
+        except Exception as e:
+            logger.debug(e)
+            logger.warn("Wayback service failed to start")
+            raise e
+
+        Wayback.wait_for_service(Wayback.PORT)
 
 
 # Runs Chromium and drives it via CDP
 class Driver(object):
 
     CHROMIUM_REVISION = 610995
+    CDP_PORT = 9222
+    PYWB_HOST = 'localhost'
+    PROXY_HOST='localhost'
+    PROXY_PORT = 8081
 
-    def __init__(self, cdp_port=9222, pywb_port=8080):
-        self.cdp_port = cdp_port
-        self.pywb_port = pywb_port
+    @classmethod
+    def new_recording_driver(cls, coll_name):
+        return cls('record', coll_name)
+
+    @classmethod
+    def new_replay_driver(cls):
+        return cls('replay')
+
+    def __init__(self, mode, coll_name=None):
+        self.mode = mode
+        if coll_name:
+            self.coll_name = coll_name
         self.chromium_executable = None
         os.environ['PYPPETEER_CHROMIUM_REVISION'] = str(self.CHROMIUM_REVISION)
         from pyppeteer import chromium_downloader
         self.chromium_downloader = chromium_downloader
         self.proc = None
-
-    def start(self):
         self.chromium_executable = self.chromium_downloader.chromium_executable()
         logger.info(self.chromium_executable)
+        self.flags = []
+        if self.mode == 'replay':
+            self.flags = [
+                '--proxy-server=http={}:{};https={}:{}'.format(self.PROXY_HOST, self.PROXY_PORT, self.PROXY_HOST, Wayback.PORT),
+                '--ignore-certificate-errors',
+                '--disk-cache-dir=/dev/null',
+                '--disk-cache-size=1'
+            ]
+        elif self.mode == 'record':
+            self.flags = [
+                '--disk-cache-dir=/dev/null',
+                '--disk-cache-size=1'
+            ]
 
+    def start(self):
         if not self.chromium_executable.exists():
             logger.info("Downloading Chromium browser")
             self.chromium_downloader.download_chromium()
 
         logger.info("Chrome Executable: {}".format(self.chromium_executable))
+
         self.proc = subprocess.Popen([
             self.chromium_executable,
-            '--remote-debugging-port={}'.format(self.cdp_port)],
-                         stdout=subprocess.DEVNULL,
-                         stderr=subprocess.DEVNULL)
+            '--remote-debugging-port={}'.format(self.CDP_PORT),
+            *self.flags
+        ], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
         tries = 5; res = None
         while tries > 0:
@@ -134,42 +182,53 @@ class Driver(object):
             raise Exception("Bad status code from Chrome: {}".format(res.status_code))
 
         logger.info("Chromium is fired up and ready to go!")
-        return self
+
 
     def cdp_url(self):
-        return "http://localhost:{}".format(self.cdp_port)
+        return "http://localhost:{}".format(self.CDP_PORT)
 
-    def kill(self):
-        self.proc.kill()
+    def stop(self):
+        if self.proc: self.proc.kill()
 
-    def record(self, entry_url, coll_name="warc-data"):
-        url_queue = deque([entry_url])
+    def replay(self, url_to_visit):
+
+        browser = pychrome.Browser(url=self.cdp_url())
+        tab = browser.new_tab()
+        tab.start()
+        tab.call_method("Network.enable")
+        tab.call_method("Page.navigate", url=url_to_visit)
+
+
+    def record(self, url_to_visit):
         browser = pychrome.Browser(url=self.cdp_url())
 
-        while len(url_queue):
-            url_to_visit = url_queue[0]
-            logger.info("Recording {}".format(url_to_visit))
-            record_url = "http://localhost:{}/{}/record/{}".format(self.pywb_port, coll_name, url_to_visit)
-            tab = browser.new_tab()
-            tab.start()
-            tab.call_method("Network.enable")
-
+        logger.info("Recording {}".format(url_to_visit))
+        record_url = "http://{}:{}/{}/record/{}".format(self.PYWB_HOST, Wayback.PORT, self.coll_name, url_to_visit)
+        tab = browser.new_tab()
+        tab.start()
+        seconds_since_something_happened = 0
+        def reset_secs(**args):
             seconds_since_something_happened = 0
-            def reset_secs(**args):
-                seconds_since_something_happened = 0
-            tab.set_listener("Network.loadingFinished", reset_secs)
-            tab.call_method("Page.navigate", url=record_url)
-            while seconds_since_something_happened < 20:
-                logger.info("Waiting for resources to load in browser")
-                tab.wait(1)
-                seconds_since_something_happened += 1
+        tab.set_listener("Network.loadingFinished", reset_secs)
+        tab.call_method("Page.navigate", url=record_url)
+        while seconds_since_something_happened < 20:
+            logger.info("Waiting for resources to load in browser")
+            tab.wait(1)
+            seconds_since_something_happened += 1
 
-            # TODO: Depending on CLI flags, pause and let
-            # the user do some manual recordriving
-            tab.stop()
-            browser.close_tab(tab)
-            url_queue.popleft()
-            return 0
+        tab.stop()
+        browser.close_tab(tab)
+        return 0
+
+subprocess_manager = SubprocessManager()
+
+def shutdown(sig, frame):
+    subprocess_manager.shutdown()
+    sys.exit(0)
+
+
+def register(stopable):
+    subprocess_manager.register(stopable)
 
 def pack_warc(rpz_file, pywb_root, coll='warc-data'):
     logger.debug("pack warc")
@@ -183,6 +242,12 @@ def pack_warc(rpz_file, pywb_root, coll='warc-data'):
     # TODO:
     # add indexes, or make sure playback includes indexing
     return 0
+
+def find_container(target):
+    unpacked_info = read_dict(target)
+    image_name = unpacked_info['current_image'].decode()
+    client = docker.from_env()
+    return list(c for c in client.containers.list() if c.image.tags.count("{}:latest".format(image_name)))[0]
 
 def run_site(args):
     rpz_file = args.pack[0]
@@ -216,6 +281,7 @@ def run_site(args):
         docker_setup(args)
         subprocess.Popen(['wb-manager', 'init', 'warc-data'], cwd=args.target[0])
 
+
     args.__setattr__('detach', True)
     args.__setattr__('expose_port', ['{}:{}'.format(args.port, args.port)])
     args.__setattr__('x11', None)
@@ -228,9 +294,9 @@ def run_site(args):
     if not args.skip_run:
         docker_run(args)
 
-    container_id = subprocess.check_output(args.docker_cmd.split() + ['ps', '-l', '-q']).decode('ascii')[:-1]
+    logger.debug("Finding reprounzip container")
+    container = find_container(target)
 
-    container_name = json.loads(subprocess.check_output(args.docker_cmd.split() + ['container', 'inspect', container_id]).decode('ascii'))[0]['Name'][1:]
 
     tries = 20
     success = False
@@ -252,31 +318,33 @@ def run_site(args):
     else:
         raise Exception("Failed to start site")
 
-    return (container_name, port)
+    return (container.name, port)
 
 
 # TODO
 # validate the warc
-# ask user if they need to do manual browse?
 # stop the reprounzip docker
 def record(args):
     error = None
+    success = False
     try:
         container_name, port = run_site(args)
-        #reprozip/reprounzip/reprounzip/unpackers/common/misc.py#474
         signal.signal(signal.SIGINT, shutdown)
 
-        recorder = Recorder(args.target[0], args.quiet).start()
-        subprocess_manager.register(recorder)
-
-        driver = Driver().start()
-        subprocess_manager.register(driver)
-
         logger.info("Start recording")
+        recorder = Wayback.new_recorder(args.target[0], args)
+        recorder.start()
+        register(recorder)
+
+        logger.info("Start browser")
+        driver = Driver.new_recording_driver('warc-data')
+        driver.start()
+        register(driver)
+
         driver.record("http://localhost:{}".format(port))
         time.sleep(5) # ensure wayback finishes writing warc
-        recorder.kill()
-        pack_warc(args.pack[0], args.target[0])
+        success = True
+
     except Exception as e:
         logger.critical("Failure to record")
         error = e
@@ -284,11 +352,75 @@ def record(args):
         subprocess_manager.shutdown()
         if error:
             raise error
+        if success:
+            pack_warc(args.pack[0], args.target[0])
         sys.exit(0)
 
 
 def playback(args):
-    print("Not implemented")
+    replay_server_name = 'rpzdj-repl.ay'
+    error, network, proxy_container = None, None, None
+    proxy_port = 8081
+    try:
+        container_name, port = run_site(args)
+        signal.signal(signal.SIGINT, shutdown)
+
+        logger.debug("Container {}".format(container_name))
+
+        client = docker.from_env()
+        logger.info("Pulling nginx container...this may take a while")
+        client.images.pull('nginx:latest')
+        logger.info("Pulling pywb container...this may take a while")
+        client.images.pull('webrecorder/pywb:latest')
+
+        network = client.networks.create(
+            "rpzdj_{}".format(time.time_ns()),
+            driver="bridge",
+            attachable=True
+        )
+        logger.info("PROXY NETWORK {}".format(network.name))
+        network.connect(container_name)
+
+        pywb_container = client.containers.run('webrecorder/pywb', detach=True, remove=True, name='pywb-playback', network=network.name, volumes={'{}/{}'.format(os.getcwd(), args.target[0]): {'bind': '/webarchive'}, '{}/pywb-playback-config.yaml'.format(os.getcwd()): {'bind': '/webarchive/config.yaml'}}, ports={'8080/tcp': Wayback.PORT})
+        register(pywb_container)
+        Wayback.wait_for_service(Wayback.PORT)
+
+        conf_string = subprocess.check_output(['sed', '-e', 's/PROXIED_SERVER/{}:{}/'.format(container_name, port), '-e', 's/SERVER_NAME/{}/'.format(replay_server_name), '-e', 's/PYWB_PORT/{}/'.format(Wayback.PORT), '-e', 's/PROXY_PORT/{}/'.format(proxy_port), 'replay-proxy-nginx.conf'])
+        conf_file = open('replay-proxy-for-{}.conf'.format(container_name), 'w')
+        conf_file.write(conf_string.decode())
+        conf_file.close()
+
+        proxy_container = client.containers.run('nginx', detach=True, remove=True, name='replay-proxy', network=network.name, volumes={'{}/replay-proxy-for-{}.conf'.format(os.getcwd(), container_name): {'bind': '/etc/nginx/conf.d/server.conf', 'mode': 'ro'}}, ports={'{}/tcp'.format(proxy_port): proxy_port})
+        register(proxy_container)
+
+        driver = Driver.new_replay_driver()
+        driver.start()
+        register(driver)
+        driver.replay("http://{}".format(replay_server_name))
+
+        input("Press Enter to quit")
+
+    except Exception as e:
+        error = e
+    finally:
+        if network:
+            network.disconnect(container_name)
+            try:
+                network.disconnect(pywb_container)
+            except docker.errors.NotFound:
+                pass
+            try:
+                network.disconnect(proxy_container)
+            except docker.errors.NotFound:
+                pass
+            network.remove()
+        subprocess_manager.shutdown()
+
+        if error:
+            raise error
+        sys.exit(0)
+
+# TODO - stop containers too
 
 
 def setup(parser, **kwargs):
@@ -310,19 +442,12 @@ def setup(parser, **kwargs):
     subparsers = parser.add_subparsers(title="actions",
                                        metavar='', help=argparse.SUPPRESS)
 
-    def add_opt_rpz_file(parser):
+    for mode in ['playback', 'record']:
+        parser = subparsers.add_parser(mode)
         parser.add_argument('pack', nargs=1, help="RPZ file")
-
-
-    parser_record = subparsers.add_parser('record')
-    add_opt_rpz_file(parser_record)
-    parser_record.add_argument('target', nargs=1, help="target directory")
-    parser_record.add_argument('--port', dest='port', help="webserver port")
-    parser_record.add_argument('--skip-setup', action='store_true', help="skip reprounzip setup")
-    parser_record.add_argument('--skip-run', action='store_true', help="skip reprounzip run")
-    parser_record.add_argument('--quiet', action='store_true', help="shhhhhhh")
-
-    parser_record.set_defaults(func=record)
-
-    parser_playback = subparsers.add_parser('playback')
-    add_opt_rpz_file(parser_playback)
+        parser.add_argument('target', nargs=1, help="target directory")
+        parser.add_argument('--port', dest='port', help="webserver port")
+        parser.add_argument('--skip-setup', action='store_true', help="skip reprounzip setup")
+        parser.add_argument('--skip-run', action='store_true', help="skip reprounzip run")
+        parser.add_argument('--quiet', action='store_true', help="shhhhhhh")
+        parser.set_defaults(func=globals()[mode])
